requirements.txt
pyshark
kafka-python
grpcio
grpcio-tools